+++ app/api/src/module/user/user.controller.ts
@@ -32,7 +32,7 @@ import { QueryParseStringPipe } from '@common/pipes/queryString.pipe'
 export class UserController {
     constructor(private readonly UserService: UserService) {}
 
-    // @UseGuards(JwtAuthGuard)
+    @UseGuards(JwtAuthGuard)
     @Get()
     @ApiOperation({
         operationId: 'getUsers',
@@ -76,6 +76,7 @@ export class UserController {
         return await this.UserService.SearchUser(search, page)
     }
 
+    @UseGuards(JwtAuthGuard)
     @Get('/:name')
     @ApiOperation({
         operationId: 'getUser',
@@ -93,11 +94,13 @@ export class UserController {
         return await this.UserService.checkUser(await this.UserService.getUser(name))
     }
 
+    @UseGuards(JwtAuthGuard)
     @Get('username/:name/')
     async GetUserByUserName(@Param('name', ParseStringPipe) name: string): Promise<UserGetDto> {
         return await this.UserService.checkUser(await this.UserService.getUserbyUserName(name))
     }
 
+    @UseGuards(JwtAuthGuard)
     @Patch('/:name')
     @ApiOperation({
         operationId: 'updateUser',
@@ -113,13 +116,17 @@ export class UserController {
     async UpdateUser(
         @Param('name', ParseStringPipe) name: string,
         @Body(new ValidationPipe()) data: UserPatchDto,
+        @Req() req,
     ): Promise<UserGetDto> {
+        if (name !== req.user.login) throw new BadRequestException('You cannot add a friend for someone else')
         const existingUser: UserGetDto = await this.UserService.getUserForPatch(name)
         return await this.UserService.updateUser(data, existingUser.login)
     }
 
+    @UseGuards(JwtAuthGuard)
     @Delete('/:name')
-    async DeleteUser(@Param('name', ParseStringPipe) name: string): Promise<UserGetDto> {
+    async DeleteUser(@Param('name', ParseStringPipe) name: string, @Req() req): Promise<UserGetDto> {
+        if (name !== req.user.login) throw new BadRequestException('You cannot add a friend for someone else')
         return await this.UserService.DeleteUser(name)
     }
 }
